* origin:
  changing examples in setrun.py  to match defaults
  modified default sizes to match original default in amr_module
  add max1d param to setrun.py in a few examples
  implemented variable max1d for 1d and 3d too
  fix proper nesting bug and reset timers
  variable max1d in amr_module can be set in setrun
  more bug fixes for timing properly after restart
  add timing to setplot.py in two examples
  fix some timing things, now works also with restarts
  Change passing aux value limits
  change all args to system_clock to integer(kind=8)
  actually use the right variable name
  update adjoint write method
  no longer using timeSetaux in module
  changes for faster filpatch to amr; already done in geo
  change status of timing file to unknown
  saving my version

+++ src/2d/filval.f90
@@ -32,7 +32,7 @@ subroutine filval(val, mitot, mjtot, dx, dy, level, time,  mic,          &
     logical :: sticksoutxfine, sticksoutyfine,sticksoutxcrse,sticksoutycrse
     
     !for setaux timing
-    integer :: clock_start, clock_finish, clock_rate
+    integer(kind=8) :: clock_start, clock_finish, clock_rate
     real(kind=8) :: cpu_start, cpu_finish
